06_django_advance/05_API_SERVER/requirements.txt
astroid==2.3.2
backcall==0.1.0
certifi==2019.9.11
chardet==3.0.4
colorama==0.4.1
coreapi==2.3.3
coreschema==0.0.4
decorator==4.4.0
Django==2.2.6
django-rest-framework==0.1.0
djangorestframework==3.10.3
drf-yasg==1.17.0
idna==2.8
inflection==0.3.1
ipython==7.8.0
ipython-genutils==0.2.0
isort==4.3.21
itypes==1.1.0
jedi==0.15.1
Jinja2==2.10.3
lazy-object-proxy==1.4.2
MarkupSafe==1.1.1
mccabe==0.6.1
packaging==19.2
parso==0.5.1
pickleshare==0.7.5
prompt-toolkit==2.0.10
Pygments==2.4.2
pylint==2.4.3
pyparsing==2.4.2
pytz==2019.3
requests==2.22.0
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.2.0
six==1.12.0
sqlparse==0.3.0
traitlets==4.3.3
typed-ast==1.4.0
uritemplate==3.0.0
urllib3==1.25.6
wcwidth==0.1.7
wrapt==1.11.2
